pyproject.toml
[project]
name = "transtructiver"
version = "0.1.0"
description = "This project delivers TranStructIVer: a modular and extensible software framework for automated semantics-preserving code transformation and verification."
readme = "README.md"
authors = [
    { name = "Loviza Sahlén", email = "losa2300@student.miun.se" },
    { name = "Thomas Springfeldt", email = "thsp1900@student.miun.se" }
]
requires-python = ">=3.14"
dependencies = [
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "pyyaml>=6.0.2",
    "tree-sitter>=0.25.2",
    "tree-sitter-language-pack>=0.13.0",
    "tree-sitter-python>=0.25.0",
]

[project.scripts]
transtructiver = "transtructiver:main"
proto-mutate = "transtructiver.prototype.mutation.mutate:main"
proto-parse = "transtructiver.prototype.parsing.parse:main"
proto-verify = "transtructiver.prototype.verification.verify:main"
proto-cli = "transtructiver.prototype.cli:main"

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
  | migrations
)/
'''

[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black==24.1.0",
    "pytest>=9.0.2",
]
